requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
pandas==2.1.3
plotly==5.18.0
gunicorn==21.2.0
supabase==2.3.4
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
